+++ Unity/HuatuoData/LocalIl2CppData/il2cpp/libil2cpp/os/Posix/Cryptography.cpp
@@ -39,7 +39,7 @@ namespace os
     {
     }
 
-    bool Cryptography::FillBufferWithRandomBytes(void* provider, intptr_t length, unsigned char* data)
+    bool Cryptography::FillBufferWithRandomBytes(void* provider, uint32_t length, unsigned char* data)
     {
         int count = 0;
         ssize_t err;
