+++ Unity/HuatuoData/LocalIl2CppData/il2cpp/libil2cpp/os/WinRT/Cryptography.cpp
@@ -27,10 +27,10 @@ namespace os
         // Do nothing, since we never allocated it
     }
 
-    bool Cryptography::FillBufferWithRandomBytes(void* provider, intptr_t length, unsigned char* data)
+    bool Cryptography::FillBufferWithRandomBytes(void* provider, uint32_t length, unsigned char* data)
     {
         NO_UNUSED_WARNING(provider);
-        return SUCCEEDED(BCryptGenRandom(NULL, data, (ULONG)length, BCRYPT_USE_SYSTEM_PREFERRED_RNG));
+        return SUCCEEDED(BCryptGenRandom(NULL, data, length, BCRYPT_USE_SYSTEM_PREFERRED_RNG));
     }
 }
 }
